packages/boot-pydbc/pyproject.toml
[project]
name = "alt-python-boot-pydbc"
version = "1.0.0"
description = "Spring Boot-style CDI auto-configuration for pydbc (relational DB)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alt-python-cdi",
    "alt-python-boot-lib",
    "alt-python-pydbc-core>=1.0.0",
    "alt-python-pydbc-sqlite>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["boot_pydbc"]

[tool.uv.sources]
alt-python-cdi = { workspace = true }
alt-python-boot-lib = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
